Qt: IP input improvements
allow empty IP inputs (which causes an empty address struct to be used).
on any erroneous IP input (excluding more than three digits in an octet, a trailing period, and anything after such period), the offended field reverts to its previous input.
also cleared a compiler warning for SocketOpen, and removed EWOULDBLOCK as an is-connecting status for *nix.